Reorganized CodeCompletions
